requirements.txt
google_api_python_client==2.147.0
protobuf==5.28.3
